+++ packages/types/custom/CHANGELOG.md
@@ -5,6 +5,10 @@
 - Feature: Add `stylePath` to `MapConfig` for a cleaner style import.
 - Feature: Add new parameter `closeIcon` to `MoveHandleProperties`.
 - Feature: Add `moveHandleActionButton` to `CoreGetters` and `CoreState` and remove it as the optional parameter `actionButton` from `MoveHandleProperties`.
+- Feature: Add new exported types `RequestGfiMode` and `GfiFeatureInformation`.
+- Fix: Add missing parameter `maxFeatures` to `GfiLayerConfiguration`.
+- Fix: Add missing parameter `mode` to `GfiConfiguration`.
+- Fix: Add missing parameter `gfiMode` to `LayerConfiguration`.
 
 ## 1.2.0